Refactor: income 관련 prisma refactoring
