+++ hps/app/incomeSource/page.tsx
@@ -2,7 +2,7 @@ import Text from '@/components/atoms/Text';
 import Title from '@/components/atoms/Title';
 import HeaderLayout from '@/components/templates/HeaderLayout';
 import { redirect } from 'next/navigation';
-import { getIncomeByUserId } from '@/lib/actions/income-select-actions';
+import { getIncomeByUserId } from '@/lib/actions/income-actions';
 import { auth } from '@/lib/auth';
 import IncomeSelectorSection from './components/IncomeSelectorSection';
