+++ hps/app/mypage/components/ProfileItem.tsx
@@ -6,7 +6,7 @@ import Text from '@/components/atoms/Text';
 import ToastMsg from '@/components/molcules/ToastMsg';
 import Image from 'next/image';
 import { useState } from 'react';
-import { updateUserField } from '@/lib/actions/users';
+import { updateUserField } from '@/lib/users';
 
 type Props = {
   id: number;
