+++ hps/constants/incomeData.ts
@@ -118,36 +118,36 @@ export const incomeData = [
     trans_memo: '외주 프로젝트',
     id: 42,
   },
-  {
-    trans_dtime: '20250602190944',
-    trans_type: '03',
-    trans_amt: 584930,
-    branch_name: '하나은행 마곡나루점',
-    trans_memo: '틱톡',
-    id: 34,
-  },
-  {
-    trans_dtime: '20250602190944',
-    trans_type: '03',
-    trans_amt: 888800,
-    branch_name: '하나은행 마포점',
-    trans_memo: '(주)디자인펀치',
-    id: 35,
-  },
-  {
-    trans_dtime: '20250615190944',
-    trans_type: '03',
-    trans_amt: 1120800,
-    branch_name: '하나은행 신촌점',
-    trans_memo: '(주)블루웨일',
-    id: 36,
-  },
-  {
-    trans_dtime: '20250612191744',
-    trans_type: '03',
-    trans_amt: 98100,
-    branch_name: '하나은행 신촌점',
-    trans_memo: '광고',
-    id: 37,
-  },
+  // {
+  //   trans_dtime: '20250602190944',
+  //   trans_type: '03',
+  //   trans_amt: 584930,
+  //   branch_name: '하나은행 마곡나루점',
+  //   trans_memo: '틱톡',
+  //   id: 34,
+  // },
+  // {
+  //   trans_dtime: '20250602190944',
+  //   trans_type: '03',
+  //   trans_amt: 888800,
+  //   branch_name: '하나은행 마포점',
+  //   trans_memo: '(주)디자인펀치',
+  //   id: 35,
+  // },
+  // {
+  //   trans_dtime: '20250615190944',
+  //   trans_type: '03',
+  //   trans_amt: 1120800,
+  //   branch_name: '하나은행 신촌점',
+  //   trans_memo: '(주)블루웨일',
+  //   id: 36,
+  // },
+  // {
+  //   trans_dtime: '20250612191744',
+  //   trans_type: '03',
+  //   trans_amt: 98100,
+  //   branch_name: '하나은행 신촌점',
+  //   trans_memo: '광고',
+  //   id: 37,
+  // },
 ];
