+++ hps/lib/actions/income-actions.ts
@@ -3,7 +3,9 @@
 import { toUtcFromSeoul } from '@/utils/time';
 import { startOfMonth, endOfMonth } from 'date-fns';
 import prisma from '../db';
+import { getLastYearSamePeriodIncomeSum } from '../income';
 
+// 해당 userId를 가진 유저의 지정된 날짜로부터 현재까지의 수입원 내역 조회
 export const getIncomesWithUserId = async (userId: number, startDate: Date) => {
   const now = new Date();
   const utcTime = toUtcFromSeoul(now.toISOString());
@@ -18,45 +20,24 @@ export const getIncomesWithUserId = async (userId: number, startDate: Date) => {
   });
 };
 
-export const getLatestSixMonthIncomesWithUserId = async (userId: number) =>
+// 6개월 치 수입액 총합 조회
+// monthAgo :  6개월 치 수입 총합을 구할 현재로부터 과거로 돌아갈 개월 수
+export const getSixMonthIncomesWithUserId = async (
+  userId: number,
+  monthsAgo: number
+) =>
   await prisma.$queryRaw<{ yearMonth: string; totalIncome: number }[]>`
   SELECT
     DATE_FORMAT(depositDate, '%Y-%m') AS yearMonth,
     SUM(amount) AS totalIncome
   FROM income
-  WHERE depositDate BETWEEN DATE_FORMAT(DATE_SUB(CURDATE(), INTERVAL 5 MONTH), '%Y-%m-01')
+  WHERE depositDate BETWEEN DATE_FORMAT(DATE_SUB(CURDATE(), INTERVAL ${monthsAgo} MONTH), '%Y-%m-01')
                       AND NOW() and userId = ${userId}
   GROUP BY yearMonth
   ORDER BY yearMonth desc;
 `;
 
-export const getLastYearSixMonthIncomesWithUserId = async (userId: number) =>
-  await prisma.$queryRaw<{ yearMonth: string; totalIncome: number }[]>`
-  SELECT
-    DATE_FORMAT(depositDate, '%Y-%m') AS yearMonth,
-    SUM(amount) AS totalIncome
-  FROM income
-  WHERE depositDate BETWEEN DATE_SUB(DATE_FORMAT(CURDATE(), '%Y-%m-01'), INTERVAL 17 MONTH)
-                      AND LAST_DAY(DATE_SUB(DATE_FORMAT(CURDATE(), '%Y-%m-01'), INTERVAL 12 MONTH)) and userId = ${userId}
-  GROUP BY yearMonth
-  ORDER BY yearMonth desc;
-`;
-
-export const getThisYearIncome = async (
-  userId: number,
-  startOfYear: Date,
-  endOfYear: Date
-) =>
-  prisma.income.findMany({
-    where: {
-      userId,
-      depositDate: {
-        gte: startOfYear,
-        lte: endOfYear,
-      },
-    },
-  });
-
+// 시작일과 종료일 사이의 수입원 내역 조회
 export const getMonthlyIncome = async (
   userId: number,
   startOfMonth: Date,
@@ -71,6 +52,8 @@ export const getMonthlyIncome = async (
       },
     },
   });
+
+// 한 달 동안의 수입 내역 조회
 export const getMonthlyIncomeWithUserId = async (
   userId: number,
   yearMonth: string
@@ -89,6 +72,7 @@ export const getMonthlyIncomeWithUserId = async (
   return incomeList;
 };
 
+// 수입 내역 수정
 export const updateIncome = async (formdate: FormData) => {
   const id = Number(formdate.get('id'));
   const incomeSource = formdate.get('value')?.toString();
@@ -99,6 +83,7 @@ export const updateIncome = async (formdate: FormData) => {
   });
 };
 
+// 해당 월의 작년 동월과의 수입액 차이 조회
 export const getIncomeChangeFromLastMonth = async (
   userId: number,
   yearMonth: string
@@ -171,64 +156,6 @@ export const getIncomeSumByPeriod = async (
   return incomes.reduce((sum, s) => sum + s.amount, 0);
 };
 
-// 지난 달 입금 합
-export const getLastMonthIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const start = new Date(now.getFullYear(), now.getMonth() - 1, 1);
-  const end = new Date(now.getFullYear(), now.getMonth(), 0, 23, 59, 59, 999);
-  return getIncomeSumByPeriod(userId, start, end);
-};
-
-// 직전 3개월 수입 합 (3, 4, 5월 등)
-export const getRecent3MonthsIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const start = new Date(now.getFullYear(), now.getMonth() - 3, 1);
-  const end = new Date(now.getFullYear(), now.getMonth(), 0, 23, 59, 59, 999);
-  return getIncomeSumByPeriod(userId, start, end);
-};
-
-// 직전 6개월 수입 합
-export const getRecent6MonthsIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const start = new Date(now.getFullYear(), now.getMonth() - 6, 1);
-  const end = new Date(now.getFullYear(), now.getMonth(), 0, 23, 59, 59, 999);
-  return getIncomeSumByPeriod(userId, start, end);
-};
-
-// 이번 달 1일부터 오늘까지 수입 합
-export const getThisMonthUntilTodayIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const utcTime = toUtcFromSeoul(now.toISOString());
-  const start = new Date(now.getFullYear(), now.getMonth(), 1);
-  return getIncomeSumByPeriod(userId, start, utcTime);
-};
-
-// 작년 동일 월 수입 합 (지금이 6월이면 2024년 6월)
-export const getLastYearSameMonthIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const year = now.getFullYear() - 1;
-  const month = now.getMonth(); // 0-indexed
-  const start = new Date(year, month, 1);
-  const end = new Date(year, month + 1, 0, 23, 59, 59, 999);
-  return getIncomeSumByPeriod(userId, start, end);
-};
-
-// 작년 동일 분기 수입 합 (지금이 6월이면 3~5월)
-export const getLastYearSamePeriodIncomeSum = async (userId: number) => {
-  const now = new Date();
-  const start = new Date(now.getFullYear() - 1, now.getMonth() - 3, 1);
-  const end = new Date(
-    now.getFullYear() - 1,
-    now.getMonth(),
-    0,
-    23,
-    59,
-    59,
-    999
-  );
-  return getIncomeSumByPeriod(userId, start, end);
-};
-
 // 작년 다음 달 수입 합 (지금이 6월이면 작년 7월)
 export const getLastYearNextMonthIncomeSum = async (userId: number) => {
   const now = new Date();
@@ -239,16 +166,87 @@ export const getLastYearNextMonthIncomeSum = async (userId: number) => {
   return getIncomeSumByPeriod(userId, start, end);
 };
 
+// 다음 달 예측 수입 조회
 export async function getPredictedNextMonthIncome(
   userId: number
 ): Promise<number> {
+  const now = new Date();
+  const threeMonthAgoStart = new Date(now.getFullYear(), now.getMonth() - 3, 1);
+  const end = new Date(now.getFullYear(), now.getMonth(), 0, 23, 59, 59, 999);
+
   const [lastYearNextMonthSum, recent3MonthsSum, lastYear3MonthsSum] =
     await Promise.all([
       getLastYearNextMonthIncomeSum(userId),
-      getRecent3MonthsIncomeSum(userId),
+      getIncomeSumByPeriod(userId, threeMonthAgoStart, end),
       getLastYearSamePeriodIncomeSum(userId),
     ]);
 
   const growthRate = recent3MonthsSum / lastYear3MonthsSum || 1;
   return lastYearNextMonthSum * growthRate;
 }
+
+// 주요 수입원 다건 저장
+export const createIncomeSources = async (
+  userId: number,
+  data: { depositorName: string; amount: number; depositDate: Date }[]
+) => {
+  return prisma.income.createMany({
+    data: data.map((item) => ({
+      userId,
+      depositorName: item.depositorName,
+      amount: item.amount,
+      depositDate: item.depositDate,
+    })),
+  });
+};
+
+// 사용자별 전체 주요 수입 내역 조회
+export const getIncomeByUserId = async (userId: number) => {
+  return prisma.income.findMany({
+    where: { userId },
+    orderBy: { depositDate: 'desc' },
+  });
+};
+
+// 수입원 삭제
+export const removeIncomeSources = async (
+  userId: number,
+  data: { depositorName: string; amount: number; depositDate: Date }[]
+) => {
+  for (const item of data) {
+    await prisma.income.deleteMany({
+      where: {
+        userId,
+        amount: item.amount,
+        depositDate: item.depositDate,
+        depositorName: item.depositorName,
+      },
+    });
+  }
+};
+
+export const getIncomeSourcesByUserId = async (
+  userId: number,
+  startDate: Date
+) => {
+  const now = new Date();
+  const utcTime = toUtcFromSeoul(now.toISOString());
+
+  return (
+    await prisma.income.groupBy({
+      by: ['incomeSource', 'depositorName'],
+      where: {
+        userId,
+        depositDate: {
+          gte: new Date(startDate),
+          lte: utcTime,
+        },
+      },
+      _sum: { amount: true },
+      orderBy: { _sum: { amount: 'desc' } },
+    })
+  ).map((item) => ({
+    category: item.incomeSource ?? item.depositorName,
+    amount: item._sum.amount ?? 0,
+  }));
+};
