+++ hps/lib/income.ts
@@ -1,6 +1,7 @@
 'use server';
 
-import prisma from '../db';
+import { getIncomeSumByPeriod } from './actions/income-actions';
+import prisma from './db';
 
 // 사용자의 아이디로 직장인 여부 확인
 export const getFirstPensionDate = async (user_id: number) => {
@@ -38,4 +39,29 @@ export const getLastIncome = async (user_id: number) => {
   });
 
   return lastIncome ? lastIncome.amount : 0; // 마지막 급여가 없으면 0 반환
+};
+
+// 작년 동일 월 수입 합 (지금이 6월이면 2024년 6월)
+export const getLastYearSameMonthIncomeSum = async (userId: number) => {
+  const now = new Date();
+  const year = now.getFullYear() - 1;
+  const month = now.getMonth(); // 0-indexed
+  const start = new Date(year, month, 1);
+  const end = new Date(year, month + 1, 0, 23, 59, 59, 999);
+  return getIncomeSumByPeriod(userId, start, end);
+};
+// 작년 동일 분기 수입 합 (지금이 6월이면 3~5월)
+export const getLastYearSamePeriodIncomeSum = async (userId: number) => {
+  const now = new Date();
+  const start = new Date(now.getFullYear() - 1, now.getMonth() - 3, 1);
+  const end = new Date(
+    now.getFullYear() - 1,
+    now.getMonth(),
+    0,
+    23,
+    59,
+    59,
+    999
+  );
+  return getIncomeSumByPeriod(userId, start, end);
 };
